Cope with K&R code in a ANSI world.
o Disable a boatload of warning designed to force K&R code out of
  existance.
o Back out some of the error.c changes
o implement abort()
o only define NULL if it's not defined (but it looks a lot like this
  code was cut and pasted because NULL is used before here).
o Add a.out.h from bsd 4.1c.

We not "link" and "run". But we don't write any a.out header and
and we think it's 2x the size. So there's issues.